@@ -1,100 +1,105 @@
 {
-  "$schema": "http://json-schema.org/draft-04/schema#",
   "name": "{{APP_TITLE}}",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
   "scripts": {
     "ls": "node commands-ls.js",
-    "build": "npm run build:stg",
-    "build:stg": "npm run clean && webpack --config ./webpack.config.staging.js --progress --colors",
-    "build:prod": "npm run clean && webpack --config ./webpack.config.prod.js --progress --colors",
+    "analyse": "NODE_ENV=analyse webpack --config webpack.config.js -p",
     "start": "npm run start:dev",
     "start:dev": "NODE_ENV=development webpack-dev-server --progress --colors --inline --hot",
-    "start:stg": "npm run clean && npm run build:stg && nodemon dist/server",
-    "start:prod": "npm run clean && npm run build:prod && nodemon dist/server",
-    "clean": "rm -rf dist && rm -rf zip",
-    "serve": "node server",
+    "start:stg": "NODE_ENV=staging webpack-dev-server --colors --inline --hot",
+    "start:prod": "NODE_ENV=production webpack-dev-server --colors --inline --hot",
+    "build": "npm run build:prod",
+    "build:stg": "npm run clean && NODE_ENV=staging webpack --config ./webpack.config.js --colors",
+    "build:prod": "npm run clean && NODE_ENV=production webpack --config ./webpack.config.js --colors",
+    "serve": "NODE_ENV=production nodemon source/server/",
+    "serve:stg": "NODE_ENV=staging nodemon source/server/",
+    "clean": "rm -rf dist",
     "test": "jest",
     "test:w": "jest --watchAll",
     "test:v": "jest --verbose",
     "test:c": "jest --coverage",
     "test:d": "jest --debug"
   },
-  "keywords": [],
   "author": "reginaldoMorais <reginaldo.cmorais@gmail.com>",
   "license": "ISC",
   "dependencies": {
-    "animate.css": "^3.5.2",
-    "axios": "^0.17.1",
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.0",
-    "babel-jest": "^22.4.1",
-    "babel-loader": "^7.1.2",
-    "babel-plugin-react-html-attrs": "^2.1.0",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-env": "^1.6.1",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "base-64": "^0.1.0",
-    "bootstrap": "4.0.0-beta.3",
-    "colors": "^1.1.2",
-    "connect-history-api-fallback": "^1.5.0",
-    "css-loader": "^0.28.8",
-    "currency-formatter": "^1.3.2",
-    "express": "^4.16.2",
-    "express-healthcheck": "^0.1.0",
-    "file-loader": "^1.1.6",
+    "@fortawesome/fontawesome": "^1.1.5",
+    "@fortawesome/react-fontawesome": "^0.0.18",
+    "babel-runtime": "^6.26.0",
+    "colors": "^1.2.1",
+    "currency-formatter": "^1.4.2",
+    "express": "^4.16.3",
     "font-awesome": "^4.7.0",
     "format-date": "^1.0.0",
-    "lodash": "^4.17.4",
-    "newrelic": "^2.6.0",
-    "path": "^0.12.7",
-    "querystring": "^0.2.0",
-    "react": "^16.2.0",
-    "react-bootstrap": "^0.32.0",
-    "react-burger-menu": "^2.2.3",
-    "react-chartjs-2": "^2.6.4",
-    "react-copy-to-clipboard": "^5.0.1",
-    "react-dom": "^16.2.0",
-    "react-redux": "^5.0.6",
-    "react-redux-toastr": "^7.2.0",
-    "react-router": "^4.2.0",
+    "loadable-components": "^1.2.0",
+    "nodemon": "^1.17.3",
+    "prop-types": "^15.6.1",
+    "react": "^16.3.0",
+    "react-burger-menu": "^2.4.1",
+    "react-dom": "^16.3.0",
+    "react-redux": "^5.0.7",
+    "react-redux-toastr": "^7.2.3",
     "react-router-dom": "^4.2.2",
-    "react-table": "^6.7.6",
-    "reactstrap": "^5.0.0-alpha.4",
+    "react-table": "^6.8.0",
+    "reactstrap": "^5.0.0",
     "redux": "^3.7.2",
     "redux-burger-menu": "^0.2.4",
-    "redux-form": "^7.2.0",
+    "redux-form": "^7.3.0",
     "redux-multi": "^0.1.12",
     "redux-persist": "^5.9.1",
     "redux-promise": "^0.5.3",
     "redux-thunk": "^2.2.0",
-    "sass-loader": "^6.0.6",
-    "serve-favicon": "^2.4.5",
-    "style-loader": "^0.19.1",
-    "sweetalert-react": "^0.4.11",
-    "yarn": "^1.3.2"
+    "sweetalert-react": "^0.4.11"
   },
   "devDependencies": {
-    "copy-webpack-plugin": "^4.3.1",
+    "axios": "^0.18.0",
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.0",
+    "babel-eslint": "^8.2.2",
+    "babel-jest": "^22.4.3",
+    "babel-loader": "^7.1.4",
+    "babel-plugin-dynamic-import-webpack": "^1.0.2",
+    "babel-plugin-syntax-dynamic-import": "^6.18.0",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-preset-env": "^1.6.1",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "bootstrap": "^4.0.0",
+    "css-loader": "^0.28.11",
     "enzyme": "^3.3.0",
     "enzyme-adapter-react-16": "^1.1.1",
-    "eslint": "^4.15.0",
+    "eslint": "^4.19.1",
+    "eslint-config-airbnb": "^16.1.0",
     "eslint-config-google": "^0.9.1",
-    "eslint-plugin-prettier": "^2.4.0",
-    "eslint-plugin-react": "^7.5.1",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "jest": "^22.4.2",
+    "eslint-plugin-import": "^2.10.0",
+    "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-prettier": "^2.6.0",
+    "eslint-plugin-react": "^7.7.0",
+    "extract-text-webpack-plugin": "^4.0.0-beta.0",
+    "file-loader": "^1.1.11",
+    "html-webpack-plugin": "^3.1.0",
+    "jest": "^22.4.3",
+    "lodash": "^4.17.5",
     "nock": "^9.2.3",
-    "node-sass": "^4.7.2",
-    "prettier": "^1.10.2",
-    "react-test-renderer": "^16.2.0",
+    "node-sass": "^4.8.3",
+    "path": "^0.12.7",
+    "prettier": "^1.11.1",
+    "progress-bar-webpack-plugin": "^1.11.0",
+    "querystring": "^0.2.0",
+    "react-copy-to-clipboard": "^5.0.1",
+    "react-test-renderer": "^16.3.1",
     "sass-lint": "^1.12.1",
+    "sass-loader": "^6.0.7",
+    "style-loader": "^0.20.3",
     "supertest": "^3.0.0",
-    "webpack": "^3.10.0",
-    "webpack-dev-server": "^2.10.1",
-    "webpack-hot-middleware": "^2.21.0",
-    "zip-webpack-plugin": "^2.1.0"
+    "uglifyjs-webpack-plugin": "^1.2.4",
+    "webpack": "^4.4.1",
+    "webpack-bundle-analyzer": "^2.11.1",
+    "webpack-cli": "^2.0.13",
+    "webpack-dev-server": "^3.1.1",
+    "webpack-node-externals": "^1.6.0"
   }
 }
