@@ -1,5 +1,5 @@
 require('babel-core/register')({});
-require('newrelic');
+require('./excludesExtensions');
 
 const PORT = process.env.PORT || 8080;
 const NODE_ENV = process.env.NODE_ENV || 'development';
